+++ desktop/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.7.9 (Unreleased)
 
 - Faster and more stable thumbnail generation.
+- Support `.supplemental-metadata` JSON files in Google Takeout.
 - .
 
 ## v1.7.8